Move canonicalization to each op
