+++ src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(Stick)
+add_subdirectory(Unstick)
