+++ src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/ShapeHelper.hpp
@@ -2,7 +2,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-//===----------------ZHighShapeHelper.hpp - help for shapes ---------------===//
+//===----------------ShapeHelper.hpp - shape helpers for ZHigh ------------===//
 //
 // Copyright 2019-2022 The IBM Research Authors.
 //
@@ -19,8 +19,8 @@
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/Value.h"
 
-#include "src/Accelerators/NNPA/Dialect/ZHigh/ZHighHelper.hpp"
 #include "src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps.hpp"
+#include "src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/OpHelper.hpp"
 #include "src/Accelerators/NNPA/Support/LayoutHelper.hpp"
 #include "src/Dialect/Mlir/IndexExpr.hpp"
 #include "src/Dialect/ONNX/ONNXOps/ShapeHelper.hpp"
