+++ src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/Stick/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_onnx_mlir_rewriter(ZHighStick)
