+++ src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/Stick/ZHighStick.td
@@ -1,0 +1,83 @@
+// SPDX-License-Identifier: Apache-2.0
+
+//===------- Stick.td - Pattern Match for ZHighStick ----------------------===//
+//
+// Copyright 2019-2022 The IBM Research Authors.
+//
+// =============================================================================
+//
+// Defines language-specific pattern match optimizations for ZHigh using
+// Declarative Rewrite Rules (DRR) specified using TableGen records.
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef STICK_TD 
+#define STICK_TD 
+
+#ifndef OP_BASE
+include "src/Accelerators/NNPA/Dialect/ZHigh/ZHigh.td"
+include "src/Dialect/ONNX/ONNX.td"
+#endif // OP_BASE
+
+include "src/Accelerators/NNPA/Dialect/ZHigh/ZHighOps/OpHelper.td"
+
+/// Note: The DRR definition used for defining patterns is shown below:
+///
+/// class Pattern<
+///    dag sourcePattern, list<dag> resultPatterns,
+///    list<dag> additionalConstraints = [],
+///    dag benefitsAdded = (addBenefit 0)
+/// >;
+
+//===----------------------------------------------------------------------===//
+// DRR patterns 
+//===----------------------------------------------------------------------===//
+
+def NoneTypeStickRemovalPattern : Pat<
+  (ZHighStickOp:$stick $arg, $layout1),
+  (replaceWithValue $arg),
+  [(IsNoneType:$arg)]
+>;
+
+// zhigh.Stick (zhigh.Unstick (%X)) = %X
+def StickUnstickSameLayoutRemovalPattern : Pat<
+  (ZHighStickOp:$stick (ZHighUnstickOp:$unstick $arg), $_),
+  (replaceWithValue $arg),
+  [(SameLayout $arg, $stick)]
+>;
+
+// zhigh.Stick (zhigh.Unstick (%X)) = onnx.LayoutTransform(%X)
+// Does not support NHWC layout because onnx.LayoutTransform requires that
+// the input and output must have the same shape, but NHWC stickify/unstickify
+// transposes the shape.
+def StickUnstickDiffLayoutRemovalPattern : Pat<
+  (ZHighStickOp:$stick (ZHighUnstickOp:$unstick $arg), $_),
+  (ONNXLayoutTransformOp $arg, (GetEncodingAttr $stick)),
+  [(NotSameLayout $arg, $stick), (NoOneIsOfNHWCLayout $arg, $stick)]
+>;
+
+// The pattern
+//   zhigh.Stick (onnx.LeakyRelu (zhigh.Unstick (%X)))
+// can be replaced by
+//   zhigh.Sub (zhigh.Relu(%X),
+//              zhigh.Relu(zhigh.Mul(%X, MinusBcastConst(%alpha))))
+//
+// Constraints:
+//   - %X should have static shape, and %alpha should be constant.
+//
+def ReplaceONNXLeakyReluPattern: Pat<
+  (ZHighStickOp:$stickout (ONNXLeakyReluOp:$out (ZHighUnstickOp $X), $alpha),
+                $layout),
+  (ZHighSubOp
+     (ZHighReluOp $X, (returnType $X)),
+     (ZHighReluOp (ZHighMulOp $X,
+                              (ZHighStickOp (GetMinusBcastConst $alpha,
+                                             $out),
+                                            $layout),
+                              (returnType $X)),
+                  (returnType $X))),
+  [(IsStaticShapeTensor $X), (IsPlusConstantFloat $alpha),
+   (SameLayout $X, $stickout)]
+>;
+
+#endif // STICK_TD
